components name & import refacto
