+++ src/cards/climates/editor.js
@@ -1,7 +1,5 @@
 import {html, nothing} from 'lit';
 
-import '../../components/button.js';
-import '../../components/input.js';
 import {
   HASS_CLIMATE_PRESET_MODE_BOOST,
   HASS_CLIMATE_PRESET_MODE_COMFORT,
