+++ src/cards/stove/editor.js
@@ -1,6 +1,4 @@
 import {html, nothing} from 'lit';
-
-import '../../components/input.js';
 import {getIcon} from '../../helpers/icons/icons.js';
 import common_style from '../../helpers/styles/common_style.js';
 import editor_common_style from '../../helpers/styles/editor_common_style.js';
