+++ src/components/inputs/inputs.js
@@ -1,0 +1,8 @@
+import './sf-input.js';
+import './sf-input-chips.js';
+import './sf-input-color-picker.js';
+import './sf-input-dropdown.js';
+import './sf-input-dropdown-entity.js';
+import './sf-input-dropdown-icon.js';
+import './sf-input-dropdown-multi-entity.js';
+import './sf-input-slider.js';
