+++ src/components/sf-accordion.js
@@ -185,10 +185,8 @@ export class SciFiAccordionCard extends LitElement {
   }
 }
 
-// Define elements
-const elements = {
-  'sci-fi-accordion-card': SciFiAccordionCard,
-};
-Object.entries(elements).forEach(([key, value]) => {
-  window.customElements.get(key) || window.customElements.define(key, value);
-});
+window.customElements.get('sci-fi-accordion-card') ||
+  window.customElements.define(
+    'sci-fi-accordion-card',
+    SciFiAccordionCard
+  );
